+++ common/src/main/java/com/buuz135/findme/network/PullItemRequestMessage.java
@@ -8,27 +8,26 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.phys.AABB;
 
-import java.util.Objects;
 
 public class PullItemRequestMessage implements CustomPacketPayload {
 
-    public static CustomPacketPayload.Type<PullItemRequestMessage> TYPE = new CustomPacketPayload.Type<>(ResourceLocation.fromNamespaceAndPath(FindMeMod.MOD_ID, "pull_item_request"));
+    public static CustomPacketPayload.Type<PullItemRequestMessage> TYPE = new CustomPacketPayload.Type<>(Identifier.fromNamespaceAndPath(FindMeMod.MOD_ID, "pull_item_request"));
     public static StreamCodec<? super RegistryFriendlyByteBuf, PullItemRequestMessage> CODEC = new StreamCodec<>() {
         @Override
         public PullItemRequestMessage decode(RegistryFriendlyByteBuf object) {
-            return new PullItemRequestMessage(ItemStack.parseOptional(object.registryAccess(), Objects.requireNonNull(object.readNbt())), object.readInt());
+            return new PullItemRequestMessage(ItemStack.OPTIONAL_STREAM_CODEC.decode(object), object.readInt());
         }
 
         @Override
         public void encode(RegistryFriendlyByteBuf registryFriendlyByteBuf, PullItemRequestMessage positionRequestMessage) {
-            registryFriendlyByteBuf.writeNbt(positionRequestMessage.stack.saveOptional(registryFriendlyByteBuf.registryAccess()));
+            ItemStack.OPTIONAL_STREAM_CODEC.encode(registryFriendlyByteBuf, positionRequestMessage.stack);
             registryFriendlyByteBuf.writeInt(positionRequestMessage.amount);
         }
     };
@@ -74,7 +73,7 @@ public class PullItemRequestMessage implements CustomPacketPayload {
                 var player = contextSupplier.getPlayer();
                 var level = player.level();
                 level.playSound(null, player.getX(), player.getY() + 0.5, player.getZ(),
-                        SoundEvents.FIRE_EXTINGUISH, SoundSource.PLAYERS, 0.5F, ((level.random.nextFloat() - level.random.nextFloat()) * 0.7F + 1.0F) * 2.0F);
+                        SoundEvents.FIRE_EXTINGUISH, SoundSource.PLAYERS, 0.5F, ((level.getRandom().nextFloat() - level.getRandom().nextFloat()) * 0.7F + 1.0F) * 2.0F);
             }
         });
         //contextSupplier.get().setPacketHandled(true);